requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.0
google-generativeai==0.8.4